add subject and teacher rest and service

+++ src/main/java/se/iths/rest/SubjectRest.java
@@ -4,38 +4,36 @@ import se.iths.entity.Subject;
 import se.iths.service.SubjectService;
 
 import javax.inject.Inject;
+import javax.validation.ConstraintViolationException;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.List;
 
-
-@Path("subject")
+@Path("/subject")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
 public class SubjectRest {
 
-   SubjectService subjectService;
-
     @Inject
-    public SubjectRest(SubjectService subjectService) {
-        this.subjectService = subjectService;
-    }
-
-    @Path("")
-    @GET
-    public Response getAllSubject() {
-
-        List<Subject> foundSubject = subjectService.getAllSubject();
-        return Response.ok(foundSubject).build();
-    }
-
+    SubjectService subjectService;
 
     @Path("")
     @POST
     public Response createSubject(Subject subject) {
-        subjectService.createSubject(subject);
-        return Response.ok(subject).build();
+        try {
+            subjectService.createSubject(subject);
+
+            return Response.ok(subject).build();
+        }
+
+
+        catch ( ConstraintViolationException error){
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity  ( "Insert a name").type(MediaType.TEXT_PLAIN_TYPE).build());
+        }
+
+
     }
 
     @Path("")
@@ -48,22 +46,59 @@ public class SubjectRest {
     @Path("{id}")
     @GET
     public Response findSubjectById(@PathParam("id") Long id) {
-        Subject foundSubject = subjectService.findSubjectById(id);
+        Subject foundStudent = subjectService.findSubjectById(id);
 
-        if (foundSubject == null) {
+        if (foundStudent == null) {
 
             throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
                     .entity("Subject with ID " + id + " was not found in database.").type(MediaType.TEXT_PLAIN_TYPE).build());
         }
-        return Response.ok(foundSubject).build();
+        return Response.ok(foundStudent).build();
     }
 
+    @Path("query")
+    @GET
+    public Response getAllSubject(@QueryParam("subject") String subject) {
+
+
+
+        String responseString = "Here is the list of  subject  " + subject;
+        return Response.ok(responseString).type(MediaType.TEXT_PLAIN_TYPE).build();
+
+
+    }
+
+
+    @Path("")
+    @GET
+    public Response getAllSubjects() {
+        List<Subject> foundStudent = subjectService.getAllSubject();
+        if (foundStudent == null) {
+
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity("The list of Subject is empty").type(MediaType.TEXT_PLAIN_TYPE).build());
+        }
+        return Response.ok(foundStudent).build();
+
+    }
 
     @Path("{id}")
     @DELETE
-    public Response deleteSubject(@PathParam("id") Long id) {
+    public Response deleteStudent(@PathParam("id") Long id) {
+
+        Subject foundStudent = subjectService.findSubjectById(id);
+
+        if (foundStudent == null) {
+
+            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
+                    .entity("Subject with ID " + id + " was not found in database.").type(MediaType.TEXT_PLAIN_TYPE).build());
+        }
         subjectService.deleteSubject(id);
         return Response.ok().build();
     }
+
+
+
+
 
 }
